backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg2-binary>=2.9
python-dotenv>=1.0
openai>=1.14
